@@ -2,22 +2,25 @@
 using KatieSite.Models;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 
 namespace KatieSite_Tests
 {
     public class FakeForumRepository : IForumRepository
     {
-        private List<ForumPost> posts = new List<ForumPost>();
+        private List<ForumPost> postList = new List<ForumPost>();
+
+        public IQueryable<ForumPost> Posts => throw new NotImplementedException(); // TODO: Fake Forum IQueryable
 
         public List<ForumPost> GetAllPosts()
         {
-            return posts;
+            return postList;
         }
 
         public ForumPost GetPostById(int postId)
         {
-            ForumPost post = posts.Find(p => p.PostId == postId);
+            ForumPost post = postList.Find(p => p.PostId == postId);
             return post;
         }
 
@@ -27,8 +30,8 @@ namespace KatieSite_Tests
             if (post != null)
             {
                 status = 1;
-                post.PostId = posts.Count+1;
-                posts.Add(post);
+                post.PostId = postList.Count+1;
+                postList.Add(post);
             }
             return status;
         }
